feat: separate common ErrorCode enum to domain specific
